Rename opener to file-opener
