+++ packages/desktop-gui/src/main.scss
@@ -5,4 +5,4 @@
 @import '!(styles)*/**/*';
 @import '../../../node_modules/@reach/dialog/styles.css';
 @import '../../ui-components/src/dropdown';
-@import '../../ui-components/src/opener/opener';
+@import '../../ui-components/src/file-opener/file-opener';
