+++ packages/reporter/src/main.scss
@@ -5,5 +5,5 @@
 @import 'lib/base';
 @import 'lib/tooltip';
 @import '../../../node_modules/@reach/dialog/styles.css';
-@import '../../ui-components/src/opener/opener';
+@import '../../ui-components/src/file-opener/file-opener';
 @import '!(lib)*/**/*';
