+++ packages/ui-components/src/file-opener/file-opener.scss
